fix: linter code
