+++ src/provider/linter.ts
@@ -84,3 +84,25 @@ export async function updateDiagnostics(
     collection.clear();
   }
 }
+
+export async function registerLinter(context: vscode.ExtensionContext) {
+  // 插件诊断器
+  const collection = vscode.languages.createDiagnosticCollection("alicloud-linter");
+  if (vscode.window.activeTextEditor) {
+    updateDiagnostics(vscode.window.activeTextEditor.document, collection);
+  }
+  context.subscriptions.push(
+    vscode.window.onDidChangeActiveTextEditor((editor) => {
+      if (editor) {
+        updateDiagnostics(editor.document, collection);
+      }
+    }),
+  );
+  context.subscriptions.push(
+    vscode.workspace.onDidChangeTextDocument((editor) => {
+      if (editor) {
+        updateDiagnostics(editor.document, collection);
+      }
+    }),
+  );
+}
